reorg createPlatformContext for client/browser
